added/moved tests
