+++ .gitignore
@@ -54,4 +54,3 @@ target/
 .idea/
 /docker/
 Dockerfile
-pia
